+++ src/options/index.ts
@@ -1,0 +1,4 @@
+export type Settings = {
+	size: number;
+	mode: string;
+};